Created a test version of RSA
I got really fascinated towards encryption and the fact that RSA encryption existed blew my mind, an encryption that allows for anyone to decipher but takes way to long to decipher, just like hiding in plain sight. So one day I decided to try recreate the RSA without the crazy long 4096 bit keys and instead tried generating smaller keys to see whether the encryption worked, the test is run on Python 2.7 and is not for normal use, as compared to genuine RSA encryption applications it requires probably 10x more time and processing speed to handle the generation of the keys. 
Feel free to test it however there are is no interface, just the IDE and the Shell to run it on to check results.